+++ lib/globals.dart
@@ -20,7 +20,7 @@ final GlobalKey<NavigatorState> navigatorKey = GlobalKey<NavigatorState>();
 
 const String prgName = 'easysend';
 const String progVersion = '0.2.260810';
-const int buildNumber = 64;
+const int buildNumber = 65;
 const String progAuthor = 'Eugen';
 
 const String langFile = 'assets/locales.json';